[Refactor] Extract shared reference-audio cache key helpers
Move the path-hash memo/sentinel/full-content-hash helpers from
higgs_tts/stages.py into preprocessing/cache_key.py unchanged so other
TTS paths can reuse the same key scheme. Higgs keeps
_reference_path_cache_key and _REF_PATH_HASH_MEMO as aliases bound to
the shared objects, so existing Higgs tests pass without modification.
