+++ sglang_omni/models/higgs_tts/stages.py
@@ -25,7 +25,6 @@ import base64
 import logging
 import os
 import threading
-from collections import OrderedDict
 from pathlib import Path
 from typing import Any
 
@@ -49,7 +48,16 @@ from sglang_omni.models.higgs_tts.utils import (
 from sglang_omni.models.higgs_tts.vocoder_scheduler import (
     HiggsStreamingVocoderScheduler,
 )
-from sglang_omni.preprocessing.cache_key import hash_bytes, hash_media_item
+
+# _REF_PATH_HASH_MEMO is the shared memo object, re-exported so tests can
+# reset it; the underscored alias keeps this module's historical API.
+from sglang_omni.preprocessing.cache_key import _REF_PATH_HASH_MEMO  # noqa: F401
+from sglang_omni.preprocessing.cache_key import (
+    hash_media_item,
+)
+from sglang_omni.preprocessing.cache_key import (
+    reference_path_cache_key as _reference_path_cache_key,
+)
 from sglang_omni.proto import StagePayload
 from sglang_omni.scheduling.bootstrap import create_sglang_infrastructure
 from sglang_omni.scheduling.omni_scheduler import OmniScheduler
@@ -71,90 +79,9 @@ _REF_CODE_CACHE_MAX_ITEMS = 256
 _REF_CODE_CACHE_MAX_BYTES = 256 * 1024 * 1024
 _REF_WAVEFORM_CACHE_MAX_ITEMS = 256
 _REF_WAVEFORM_CACHE_MAX_BYTES = 512 * 1024 * 1024
-_REF_PATH_HASH_MEMO_MAX_ITEMS = 1024
-_REF_PATH_HASH_SENTINEL_BYTES = 8192
-_REF_PATH_HASH_MEMO: OrderedDict[str, tuple[str, str]] = OrderedDict()
-_REF_PATH_HASH_MEMO_LOCK = threading.Lock()
 
 # Saturates near c=16 on H100/H200; higher client concurrency only queues.
 DEFAULT_MAX_CONCURRENCY = 16
-
-
-def _reference_path_hash_memo_key(path: Path) -> tuple[str, int] | None:
-    try:
-        if not path.is_file():
-            return None
-        stat_result = path.stat()
-        memo_key = (
-            f"{path.resolve()}:"
-            f"{stat_result.st_size}:"
-            f"{stat_result.st_mtime_ns}:"
-            f"{stat_result.st_ctime_ns}"
-        )
-        return memo_key, int(stat_result.st_size)
-    except OSError:
-        return None
-
-
-def _reference_path_sentinel(path: Path, file_size: int) -> str | None:
-    try:
-        chunk_size = min(_REF_PATH_HASH_SENTINEL_BYTES, file_size)
-        with path.open("rb") as f:
-            chunks = [f.read(chunk_size)]
-            if file_size > _REF_PATH_HASH_SENTINEL_BYTES:
-                middle_offset = max((file_size - chunk_size) // 2, 0)
-                f.seek(middle_offset)
-                chunks.append(f.read(chunk_size))
-            if file_size > 2 * _REF_PATH_HASH_SENTINEL_BYTES:
-                f.seek(max(file_size - chunk_size, 0))
-                chunks.append(f.read(chunk_size))
-        return hash_bytes(b"".join(chunks) + f"|size:{file_size}".encode())
-    except OSError:
-        return None
-
-
-def _get_reference_path_hash(memo_key: str, sentinel: str) -> str | None:
-    with _REF_PATH_HASH_MEMO_LOCK:
-        cached = _REF_PATH_HASH_MEMO.get(memo_key)
-        if cached is None:
-            return None
-        cached_sentinel, digest = cached
-        if cached_sentinel != sentinel:
-            _REF_PATH_HASH_MEMO.pop(memo_key, None)
-            return None
-        _REF_PATH_HASH_MEMO.move_to_end(memo_key)
-        return digest
-
-
-def _put_reference_path_hash(memo_key: str, sentinel: str, digest: str) -> None:
-    with _REF_PATH_HASH_MEMO_LOCK:
-        _REF_PATH_HASH_MEMO[memo_key] = (sentinel, digest)
-        _REF_PATH_HASH_MEMO.move_to_end(memo_key)
-        while len(_REF_PATH_HASH_MEMO) > _REF_PATH_HASH_MEMO_MAX_ITEMS:
-            _REF_PATH_HASH_MEMO.popitem(last=False)
-
-
-def _reference_path_cache_key(path_like: str | Path) -> str | None:
-    # Memoized full-content hash. The stat tuple avoids rereading stable local
-    # refs while still invalidating normal and rapid same-size replacements.
-    path = Path(str(path_like)).expanduser()
-    memo = _reference_path_hash_memo_key(path)
-    if memo is None:
-        return None
-    memo_key, file_size = memo
-    sentinel = _reference_path_sentinel(path, file_size)
-    if sentinel is None:
-        return None
-    digest = _get_reference_path_hash(memo_key, sentinel)
-    if digest is not None:
-        return f"file:{digest}"
-    try:
-        digest = hash_bytes(path.read_bytes())
-    except OSError:
-        return None
-    if _reference_path_hash_memo_key(path) == memo:
-        _put_reference_path_hash(memo_key, sentinel, digest)
-    return f"file:{digest}"
 
 
 def _reference_audio_cache_key(reference_audio: Any) -> str | None:
